Refactor ContainerDankNull
