--- src/main/java/p455w0rd/danknull/inventory/slot/SlotDankNullDock.java
@@ -1,15 +1,0 @@
-package p455w0rd.danknull.inventory.slot;
-
-import net.minecraftforge.items.IItemHandler;
-import net.minecraftforge.items.SlotItemHandler;
-
-/**
- * @author p455w0rd
- *
- */
-public class SlotDankNullDock extends SlotItemHandler {
-
-	public SlotDankNullDock(final IItemHandler handler, final int index, final int x, final int y) {
-		super(handler, index, x, y);
-	}
-}
